Change width and height to doubles

+++ src/main/java/breakout/Block.java
@@ -5,7 +5,7 @@ import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
 
 public class Block extends Sprite {
-    public Block(int width, int height, Point2D position, Color color) {
+    public Block(double width, double height, Point2D position, Color color) {
         super(new Rectangle(width, height, color),
                 position,
                 new Point2D(0, 0));
